transactions and admin changes

+++ eSunSpeedDomain/ListModel.cs
@@ -22,8 +22,8 @@ namespace eSunSpeedDomain
         public string Account { get; set; }
         public int TotalAmt { get; set; }
 
-        public int Debit { get; set; }
-        public int Credit { get; set; }
+        public decimal Debit { get; set; }
+        public decimal Credit { get; set; }
         public string Narration { get; set; }
         
         public string CreatedBy { get; set; }
